requirements.txt
beautifulsoup4==4.13.3
boto3==1.36.13
docling==2.18.0
fastapi[standard]
llama_parse==0.5.20
markitdown==0.0.1a3
pandas==2.2.3
Pillow
pymupdf
python-dotenv==1.0.1
python-multipart
Requests
streamlit
pathlib~=1.0.1
python-dotenv~=1.0.1
llama-parse~=0.5.20
uvicorn
